Stepper Motor Functionality
This commit marks a HUGE improvement to the library: Stepper Motor Functionality.
The same sketch as before can be used, and a new bluetooth keyword, stepperX, has been added. (where x is the stepper number.) So far stepper definitions occour in the library, so there is still improvement to the library to be made. Expects ULN2003 based stepper connected to 32,31,30,29 on the Teensy.